Former-commit-id: c30129dcdd3d7bc527d52d4ef2f0a8cb44f95dcb

+++ user/fangg/source.c
@@ -22,7 +22,7 @@ typedef struct {
     int srange;
     float alpha;
     int decay;
-    int trunc;
+    float trunc;
 }spara; /*source parameter*/
 /*^*/
